Added Tests for Ambiguous Cases Half/Quarter Fix
Added some tests to test for ambiguous cases. Obviously these aren't all
of them and most of them are failing at present. In addition, I added a
check for quarters to see if it is a verb ( by checking for preceeding pronoun
). I also added a cleanup method for halfs and hanging fractions like
"/2". This should fix the 'half an hour' chronic problem.