Added Ignore Option and Reformatted Tests
The ignore option has been added, such that an array of strings can be
specified to be ignored by numerizer, and it will no longer parse these
strings. Might need to make it clear in documentation what these strings
are.

In addition, the test cases have been reformatted. I was a little bit
map happy earlier and have now changed all of the unecessary maps to
assert equals. And of course, some tests have been written for the
ignore case.