Added Tests and Plural Check for Bias
Added additional tests for the bias.

In additon, the fractional conversion still occurs if there is a ordinal
with an 's' on it. Because in no apparent context is 'fourths' an
ordinal option.

Because of this, I redid some of the arrays to accomodate for this
change. I merged together the ambiguous fraction and ordinal arrays.

There are now also precalculated arrays for all ordinals, fractional,
and plural fractionals.

I think we will still need more tests for the biased options though.