Removed Loops using Hashes
Previously, there were a number of loops that could potentially cause
problems ... performance wise and just logic wise. Now that has been
simplified by the usage of hash. Instead of looping over nested arrays
through each and every possible number, you simply gsub checking for all
words, and then use the match as a key in a hash, returning you the
appropriate number. This way, you don't use gsub in a nested loop,
repeatedly scanning over the string an unnecessary amount of times.
However, this fix does change the logic a little bit, and may be slighty
more complex when dealing with edge cases, like with 'second' and
'quarter'. It is to be seen how this can be refactored.