Refactored Methods
Refactored the method names and defined a standard set of methods for
numerizing a sentence. We also added the Regexify method that takes a
list and turns it into a regex expression for you. It also handles
ignoring certain items within the list and providing a block to further
ignore certain items.