Fixed "a fourth of" case, added bias to all calls
First of all, the bias options was added to all calls within the
subclass. For numerals, ordinals, fractionals, and big prefixes. Because
in any case, it might be necessary, and was is the case of "a fourth
of".
Previously, the regex turned all 'a's prefixing a number/ordinal
into '<num>1'. This was undesired in the case of bias ordinal, because
we are explicitly stating that the ordinal is not a fractional. and so
prefixing it with a 1 makes little sense.