Removed Nested Checking Using Negative Lookbehind
Apparently Ruby does support variable size negative lookbehind. This
makes it easy to check if the previous word isn't word x. This means we
no longer have to break apart the string and do another check on the
regex when parsing. We can simply specify which preceeding options
should not work.